backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Database and ORM
supabase==2.7.4
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
bcrypt==4.1.2

# HTTP client and async support
httpx>=0.24,<0.26
aiofiles==23.2.1

# Date and time handling
python-dateutil==2.8.2

# Email functionality
emails==0.6
jinja2==3.1.3

# File upload and validation
python-magic==0.4.27
Pillow==10.2.0

# Logging and monitoring
structlog==24.1.0

# Testing (optional)
pytest==8.0.1
pytest-asyncio==0.23.5

# Payment processing (Stripe)
stripe==8.7.0

# Calendar integration
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utilities
python-slugify==8.0.4
